Shift experiment 2 from main to separate file
